+++ src/main/java/com/yapp/pet/domain/club/repository/elasticsearch/CustomClubSearchRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.yapp.pet.domain.club.repository;
+package com.yapp.pet.domain.club.repository.elasticsearch;
 
 import com.yapp.pet.domain.club.document.ClubDocument;
 import com.yapp.pet.domain.club.entity.EligibleSex;
