+++ src/main/java/com/yapp/pet/domain/club/service/ClubQueryService.java
@@ -4,8 +4,8 @@ import com.yapp.pet.domain.account.entity.Account;
 import com.yapp.pet.domain.club.document.ClubDocument;
 import com.yapp.pet.domain.club.entity.Club;
 import com.yapp.pet.domain.club.repository.ClubFindCondition;
-import com.yapp.pet.domain.club.repository.ClubRepository;
-import com.yapp.pet.domain.club.repository.ClubSearchRepository;
+import com.yapp.pet.domain.club.repository.jpa.ClubRepository;
+import com.yapp.pet.domain.club.repository.elasticsearch.ClubSearchRepository;
 import com.yapp.pet.domain.comment.service.CommentQueryService;
 import com.yapp.pet.global.mapper.ClubMapper;
 import com.yapp.pet.web.club.model.ClubFindDetailResponse;
