+++ src/main/java/com/yapp/pet/domain/club/service/ClubService.java
@@ -5,8 +5,8 @@ import com.yapp.pet.domain.accountclub.AccountClub;
 import com.yapp.pet.domain.accountclub.AccountClubRepository;
 import com.yapp.pet.domain.club.document.ClubDocument;
 import com.yapp.pet.domain.club.entity.Club;
-import com.yapp.pet.domain.club.repository.ClubRepository;
-import com.yapp.pet.domain.club.repository.ClubSearchRepository;
+import com.yapp.pet.domain.club.repository.jpa.ClubRepository;
+import com.yapp.pet.domain.club.repository.elasticsearch.ClubSearchRepository;
 import com.yapp.pet.domain.comment.CommentRepository;
 import com.yapp.pet.domain.pet.entity.Pet;
 import com.yapp.pet.domain.pet.repository.PetRepository;
